% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter_conditional.R
\name{conditional_particle_filter}
\alias{conditional_particle_filter}
\title{conditional_particle_filter}
\usage{
conditional_particle_filter(nparticles, model, theta, observations, randomness,
  TreeClass, ref_trajectory, with_as = FALSE)
}
\description{
runs a conditional particle filter, with or without ancestor sampling
}

