% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter_coupled_conditional.R
\name{coupled_conditional_particle_filter}
\alias{coupled_conditional_particle_filter}
\title{coupled_conditional_particle_filter}
\usage{
coupled_conditional_particle_filter(nparticles, model, theta, observations,
  randomness, ref_trajectory1, ref_trajectory2, coupled_resampling,
  final_resampling, resampling_parameters, with_as, TreeClass)
}
\description{
runs two coupled conditional particle filters, with or without ancestor sampling
}

