% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_particle_filter.R
\name{coupled_particle_filter}
\alias{coupled_particle_filter}
\title{coupled_particle_filter}
\usage{
coupled_particle_filter(nparticles, model, theta1, theta2, observations,
  randomness, coupled_resampling, resampling_parameters = list())
}
\description{
coupled_particle_filter
}

