% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter_coupled_given.R
\name{coupled_pf_given}
\alias{coupled_pf_given}
\title{coupled_pf_given}
\usage{
coupled_pf_given(nparticles, model, theta2, observations, randomness,
  coupled_resampling_given, resampling_parameters = list(), system_ref)
}
\description{
runs a coupled particle filter, given the results of another one
}

