% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoupledCPF-package.R
\docType{package}
\name{CoupledCPF-package}
\alias{CoupledCPF}
\alias{CoupledCPF-package}
\title{CoupledCPF}
\description{
...
}
\details{
...
}
\author{
Pierre E. Jacob <pierre.jacob.work@gmail.com>
}
\keyword{package}

