% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoupledPF-package.R
\docType{package}
\name{CoupledPF-package}
\alias{CoupledPF}
\alias{CoupledPF-package}
\title{CoupledPF}
\description{
...
}
\details{
...
}
\author{
Pierre E. Jacob <pierre.jacob.work@gmail.com>
}
\keyword{package}

