% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPF.R
\name{CPF}
\alias{CPF}
\title{Conditional Particle Filter}
\usage{
CPF(nparticles, model, theta, observations, ref_trajectory, with_as = FALSE)
}
\arguments{
\item{nparticles}{number of particles}

\item{model}{a list representing a model, for instance as given by \code{\link{get_ar}}.}

\item{theta}{a parameter to give to the model functions}

\item{observations}{a matrix of observations of size datalength x dimension(observation)}

\item{ref_trajectory}{a reference trajectory, of size dimension(process) x datalength; if missing, runs a standard
particle filter.}

\item{with_as}{whether ancestor sampling should be used (TRUE/FALSE)}
}
\value{
A new trajectory.
}
\description{
Runs a conditional particle filter, with or without ancestor sampling
}

