% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPF_coupled.R
\name{CPF_coupled_RB}
\alias{CPF_coupled_RB}
\title{Coupled Conditional Particle Filter}
\usage{
CPF_coupled_RB(nparticles, model, theta, observations, ref_trajectory1,
  ref_trajectory2, coupled_resampling, with_as = FALSE,
  h = function(trajectory) {     return(trajectory) })
}
\arguments{
\item{nparticles}{number of particles}

\item{model}{a list representing a model, for instance as given by \code{\link{get_ar}}.}

\item{theta}{a parameter to give to the model functions}

\item{observations}{a matrix of observations of size datalength x dimension(observation)}

\item{ref_trajectory1}{a first reference trajectory, of size dimension(process) x datalength}

\item{ref_trajectory2}{a second reference trajectory, of size dimension(process) x datalength}

\item{coupled_resampling}{a coupled resampling scheme, such as \code{\link{CR_indexmatching}}.}

\item{with_as}{whether ancestor sampling should be used (TRUE/FALSE)}

\item{h}{test function, which we want to integrate with respect to the smoothing distribution}
}
\value{
A list with 'new_trajectory1', 'new_trajectory2' and 'estimate1', 'estimate2'
}
\description{
Runs a coupled conditional particle filter, with or without ancestor sampling
}
