% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPF.R
\name{CPF_RB}
\alias{CPF_RB}
\title{Conditional Particle Filter with RB}
\usage{
CPF_RB(nparticles, model, theta, observations, ref_trajectory = NULL,
  with_as = FALSE, h = function(trajectory) {     return(trajectory) })
}
\arguments{
\item{nparticles}{number of particles}

\item{model}{a list representing a model, for instance as given by \code{\link{get_ar}}.}

\item{theta}{a parameter to give to the model functions}

\item{observations}{a matrix of observations of size datalength x dimension(observation)}

\item{ref_trajectory}{a reference trajectory, of size dimension(process) x datalength; if missing, runs a standard
particle filter.}

\item{with_as}{whether ancestor sampling should be used (TRUE/FALSE)}

\item{h}{test function, which we want to integrate with respect to the smoothing distribution}
}
\value{
A list with 'new_trajectory' and 'estimate'
}
\description{
Runs a conditional particle filter, with or without ancestor sampling,
and with Rao-Blackwellization
}
