% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupledresampling-hilbert.R
\name{CR_hilbert}
\alias{CR_hilbert}
\title{Coupled Resampling: Hilbert}
\usage{
CR_hilbert(xparticles1, xparticles2, normweights1, normweights2, ...)
}
\value{
Two vectors of ancestors, column-binded in a matrix.
}
\description{
This function performs coupled resampling based on sorting the samples
according to their projection on the Hilbert space filling curve.
}

