% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupledresampling-multinomial.R
\name{CR_multinomial}
\alias{CR_multinomial}
\title{Coupled Resampling: multinomial}
\usage{
CR_multinomial(xparticles1, xparticles2, normweights1, normweights2, ...)
}
\value{
Two vectors of ancestors, column-binded in a matrix.
}
\description{
This function performs independent multinomial resampling on each system
}
