% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupledresampling-systematic.R
\name{CR_systematic}
\alias{CR_systematic}
\title{Coupled Resampling: systematic}
\usage{
CR_systematic(xparticles1, xparticles2, normweights1, normweights2, ...)
}
\value{
Two vectors of ancestors, column-binded in a matrix.
}
\description{
This function performs systematic resampling with the same uniform on each system.
}
