% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_levydriven.R
\name{get_levydriven}
\alias{get_levydriven}
\title{Levy-driven stochastic volatility model as in Chopin Jacob Papaspiliopoulos 2013}
\usage{
get_levydriven()
}
\value{
A list
}
\description{
This function returns a list with objects such as
* rinit, rinit_rand to sample from the initial distribution
* rtransition, rtransition_rand to sample from the transition
* dtransition to evaluate the transition density
* dmeasurement to evaluate the measurement density
* dimension, which represents the dimension of the latent process
}
