% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_pz.R
\name{get_pz}
\alias{get_pz}
\title{Phytoplankton-zooplankton model as in Jones, Parslow, Murray 2010}
\usage{
get_pz()
}
\value{
A list
}
\description{
This function returns a list with objects such as
* rinit, rinit_rand to sample from the initial distribution
* rtransition, rtransition_rand to sample from the transition
* dtransition to evaluate the transition density
* dmeasurement to evaluate the measurement density
* dimension, which represents the dimension of the latent process
}
