% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_stochvol.R
\name{get_stochvol}
\alias{get_stochvol}
\title{Multivariate stochastic volatility model as in Chopin Gerber 2015}
\usage{
get_stochvol(dimension)
}
\value{
A list
}
\description{
This function returns a list with objects such as
* rinit, rinit_rand to sample from the initial distribution
* rtransition, rtransition_rand to sample from the transition
* dtransition to evaluate the transition density
* dmeasurement to evaluate the measurement density
* dimension, which represents the dimension of the latent process
}

