% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lorenz_transition.R
\name{lorenz_transition}
\alias{lorenz_transition}
\title{lorenz_transition}
\usage{
lorenz_transition(xparticles, time_start, time_end, dt, parameters)
}
\description{
Solve lorenz ODE for each particle, given each alpha, from time to time + 1,
and given the parameters (c, e, ml, mq).
}
