% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particlefilter.R
\name{particlefilter}
\alias{particlefilter}
\title{Particle Filter}
\usage{
particlefilter(nparticles, model, theta, observations)
}
\arguments{
\item{nparticles}{a number of particles}

\item{model}{a list representing a model, for instance as given by \code{\link{get_ar}}.}

\item{theta}{a parameter to give to the model functions}

\item{observations}{a matrix of observations of size datalength x dimension(observation)}
}
\value{
A list with all the trajectories, stored in an array of dimension dimx x (T+1) x N,
and the associated normalized weights
}
\description{
Particle filter with storage of trajectories
}

