% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rheeglynn_estimator.R
\name{rheeglynn_estimator}
\alias{rheeglynn_estimator}
\title{Rhee--Glynn smoothing estimator}
\usage{
rheeglynn_estimator(observations, model, theta, algoparameters,
  force_niterations = 0, max_niterations = 0)
}
\arguments{
\item{observations}{a matrix of observations of size datalength x dimension(observation)}

\item{model}{a list representing a model, for instance as given by \code{\link{get_ar}}.}

\item{theta}{a parameter to give to the model functions}

\item{algoparameters}{a list containing the algorithmic parameters, that is, nparticles, coupled_resampling,
with_as and  lambda. For the first three, see \code{\link{CPF_coupled}}. Lambda refers to the parameter
of the Geometric truncation variable. Set to zero to by-pass this choice.}

\item{force_niterations}{forces the algorithm to run for a set number of iterations (default to zero, which disables this feature)}

\item{max_niterations}{makes the algorithm stop at the set number of iterations (default to zero, which disables this features)}
}
\value{
a list with the estimator of the smoothing mean, the number of iterations it took, and the truncation variable if any
}
\description{
Estimates the mean of the smoothing distribution at all times.
}

