% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rheeglynn_estimator.R
\name{unbiasedestimator}
\alias{unbiasedestimator}
\title{Rhee--Glynn smoothing estimator}
\usage{
unbiasedestimator(single_kernel, coupled_kernel, rinit, h = function(x)
  x, k = 0, m = 1, max_iterations = Inf)
}
\arguments{
\item{single_kernel}{function to sample the Markov chain}

\item{coupled_kernel}{function to sample the coupled Markov chain}

\item{rinit}{function to sample from the initial distribution}

\item{h}{test function to estimate (default to h(x) = x)}

\item{k}{iteration at which to start averaging (default to 0)}

\item{m}{iteration at which to stop averaging (default to 1)}

\item{max_niterations}{iteration at which to stop the while loop (default to Infinity)}
}
\value{
a list with the estimator of the smoothing expectation (uestimator) and the meeting time (meetingtime)
}
\description{
Estimates the expectation of a function h with respect to the invariant distribution of a Markov chain
}
