% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupledresampling-indexmatching.R
\name{CR_indexmatching}
\alias{CR_indexmatching}
\title{Coupled Resampling: index-matching}
\usage{
CR_indexmatching(xparticles1, xparticles2, normweights1, normweights2, ...)
}
\value{
Two vectors of ancestors, column-binded in a matrix.
}
\description{
This function performs coupled resampling based on index-matching.
}

